requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=2.2.0=py38haa244fe_0
argon2-cffi=20.1.0=py38h294d835_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=2.106=mkl
bleach=3.3.0=pyh44b312d_0
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2020.12.5=h5b45459_0
certifi=2020.12.5=py38haa244fe_1
cffi=1.14.5=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=11.1.1=heb2d755_7
cycler=0.10.0=py_2
decorator=5.0.7=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
deprecation=2.1.0=pyh9f0ad1d_0
entrypoints=0.3=pyhd8ed1ab_1003
freetype=2.10.4=h546665d_1
icu=58.2=ha925a31_3
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.0.1=py38haa244fe_0
intel-openmp=2021.2.0=h57928b3_616
ipykernel=5.5.3=py38h43734a8_0
ipython=7.22.0=py38h43734a8_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py38haa244fe_2
jinja2=2.11.3=pyh44b312d_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter-packaging=0.9.2=pyhd8ed1ab_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38haa244fe_0
jupyter_server=1.6.4=py38haa244fe_0
jupyterlab=3.0.14=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.4.0=pyhd8ed1ab_0
kiwisolver=1.3.1=py38hbd9d945_1
libblas=3.9.0=6_mkl
libcblas=3.9.0=6_mkl
libflac=1.3.3=h0e60522_1
liblapack=3.9.0=6_mkl
liblapacke=3.9.0=6_mkl
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.37=h1d00b33_2
libsndfile=1.0.31=h0e60522_1
libsodium=1.0.18=h8d14728_1
libtiff=4.1.0=h56a325e_1
libuv=1.41.0=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
llvmlite=0.36.0=py38h57a6900_0
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38h294d835_3
matplotlib=3.4.1=py38haa244fe_0
matplotlib-base=3.4.1=py38heae8d8c_0
mistune=0.8.4=py38h294d835_1003
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
nbclassic=0.2.7=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py38haa244fe_3
nbformat=5.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=h5362a0b_0
notebook=6.3.0=pyha770c72_1
numba=0.53.1=py38hd0eb412_0
numpy=1.20.2=py38h09042cb_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1k=h8ffe710_0
packaging=20.9=pyh44b312d_0
pandoc=2.13=h8ffe710_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pesq=0.0.2=pypi_0
pickleshare=0.7.5=py_1003
pillow=8.2.0=py38h4fa10fc_0
pip=21.1=pyhd8ed1ab_0
prometheus_client=0.10.1=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
pycparser=2.20=pyh9f0ad1d_2
pydub=0.25.1=pyhd8ed1ab_0
pygments=2.8.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pypesq=1.2.4=pypi_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38h294d835_2
pysocks=1.7.1=py38haa244fe_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
pystoi=0.3.3=pypi_0
python=3.8.8=h7840368_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.8.1=py3.8_cuda11.1_cudnn8_0
pytz=2021.1=pyhd8ed1ab_0
pywin32=300=py38h294d835_0
pywinpty=0.5.7=py38haa244fe_1
pyzmq=22.0.3=py38h09162b1_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3deb0d_0
scipy=1.6.2=py38he847743_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py38haa244fe_3
sip=4.19.13=py38ha925a31_0
six=1.15.0=pyh9f0ad1d_0
sniffio=1.2.0=py38haa244fe_1
sqlite=3.35.5=h8ffe710_0
terminado=0.9.4=py38haa244fe_0
testpath=0.4.4=py_0
tk=8.6.10=h8ffe710_1
tomlkit=0.7.0=py38haa244fe_3
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cu111
tornado=6.1=py38h294d835_1
tqdm=4.60.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.2=pyhd3deb0d_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38haa244fe_1006
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
zeromq=4.3.4=h0e60522_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
